backend/requirements.txt
django==5.2.8
djangorestframework
django-cors-headers
Pillow
psycopg2-binary
pymongo
django-environ
djangorestframework-simplejwt
